requirements.txt
absl-py==1.4.0
aiosqlite==0.17.0
alembic==1.11.2
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
audioread==3.0.0
billiard==4.1.0
Brotli==1.0.9
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmaes==0.10.0
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
fastapi==0.101.0
filelock==3.12.2
fonttools==4.42.0
fsspec==2023.6.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h11==0.14.0
httptools==0.6.0
huggingface==0.0.1
huggingface-hub==0.16.4
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
kombu==5.3.1
lazy_loader==0.3
librosa==0.10.0.post2
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
mutagen==1.46.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
optuna==3.3.0
packaging==23.1
pandas==2.0.3
pescador==2.1.0
Pillow==10.0.0
pooch==1.6.0
prompt-toolkit==3.0.39
protobuf==4.23.4
pyannote.audio==1.1.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==1.5.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
pydub==0.25.1
pyfiglet==0.8.post1
Pygments==2.16.1
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.5
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tortoise-orm==0.19.3
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
yt-dlp==2023.7.6
